requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
anthropic==0.40.0
openai==1.57.0
httpx==0.28.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sse-starlette==2.1.3
tiktoken==0.8.0
aiofiles==24.1.0
tenacity==9.0.0
loguru==0.7.3
rich==13.9.4
